backend/ml-service/requirements.txt
requests>=2.31.0
numpy>=1.24.0
Pillow>=10.0.0
ultralytics>=8.0.0      # For YOLO models
opencv-python>=4.8.0    # For image processing

